requirements.txt
asn1crypto
Babel
Django==3.0.3
Markdown
Pillow
PyYAML
backports.ssl-match-hostname
certifi
chardet
cryptography
django-browserid
django-cities
django-classy-tags
django-common-helpers
django-countries
django-crispy-forms
django-cron
django-crontab
django-debug-toolbar
django-extensions
django-filter
django-ipware
django-missing
django-model-utils
django-pagedown
django-phonenumber-field
django-sekizai
django-user-agents
djangorestframework
djangotoolbox
docopt
enum34
facebook-sdk
facebookads
geoip2
google-api-python-client
httplib2
idna
include
ipaddress
jsonpickle
maxminddb
numpy
oauth2client
oauthlib
pandas
phonenumberslite
psycopg2
pycparser
pyOpenSSL
pyasn1
pyasn1-modules
pygeoip
pymongo
python-dateutil
python-gflags
pytz
paypalrestsdk
q
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
sqlparse
swapper
texttable
tqdm
tweepy
ua-parser
uritemplate
urllib3
user-agents
websocket-client
raven
django-cors-headers
gunicorn
awesome-slugify
easy_thumbnails
django-autoslug
html2text
virtualenv
redis
django-redis
django-redis-cache
django-rq
django-rq-scheduler
django_cron
django-crontab
sentry-sdk==0.14.2
mammoth
django-bot-crawler-blocker
